Move TestForm into pages directory
