+++ react-app/src/components/TestForm/TestComplete.js
@@ -1,6 +1,6 @@
 export default function TestComplete() {
   return (
-    <div className='site__sub-section terms-of-use'>
+    <div className='site__sub-section flex-dir-col'>
       <h1 className='primary-title test-complete__title'>Complete</h1>
       <p className='test-complete__thank-you'>
         Thank you for completing this test. You may now close the browser.
